requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.1
numpy==1.24.3
tensorflow==2.13.0
h5py==3.9.0
matplotlib==3.7.2
keras==2.13.1
opencv-python==4.8.1.78
